Refactor Usertimeline and Index class
Refactored the Usertimeline and Index class code such that their identical html code is in a separate class TimelinePartial and they inherit from the same timelinemodel
