+++ src/Chirp.Authentication/Pages/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/"
-@model IndexModel
+@model TimelineModel
 @{
     ViewData["Title"] = "Chirp!";
     Layout = "Shared/_Layout";
@@ -8,25 +8,5 @@
 <div>
     <h2> Public Timeline </h2>
 
-    @if (Model.Cheeps!.Any())
-    {
-        <ul id="messagelist" class="cheeps">
-            @foreach (var cheep in Model.Cheeps!)
-            {
-                <li>
-                    <p>
-                        <strong>
-                            <a href="/@cheep.Author">@cheep.Author</a>
-                        </strong>
-                        @cheep.Message
-                        <small>&mdash; @cheep.TimeStamp</small>
-                    </p>
-                </li>
-            }
-        </ul>
-    }
-    else
-    {
-        <em>There are no cheeps so far.</em>
-    }
+    <partial name="_TimelinePartial" />
 </div>
