--- src/Chirp.Authentication/Pages/UserProfile.cshtml
@@ -1,2 +1,0 @@
-@page "/findudafnoget"
-@model UserProfileModel
