+++ src/Chirp.Authentication/Pages/UserTimeline.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/author/{author}"
-@model Chirp.Authentication.Pages.UserTimelineModel
+@model Chirp.Authentication.Pages.TimelineModel
 @{
     ViewData["Title"] = "Chirp!";
     Layout = "Shared/_Layout";
@@ -9,25 +9,5 @@
 <div>
     <h2> @routeName's Timeline </h2>
 
-    @if (Model.Cheeps!.Any())
-    {
-        <ul id="messagelist" class="cheeps">
-            @foreach (var cheep in Model.Cheeps!)
-            {
-                <li>
-                    <p>
-                        <strong>
-                            <a href="/@cheep.Author">@cheep.Author</a>
-                        </strong>
-                        @cheep.Message
-                        <small>&mdash; @cheep.TimeStamp</small>
-                    </p>
-                </li>
-            }
-        </ul>
-    }
-    else
-    {
-        <em>There are no cheeps so far.</em>
-    }
+   <partial name="_TimelinePartial" />
 </div>
