+++ src/Chirp.Authentication/Pages/_TimelinePartial.cshtml
@@ -1,0 +1,23 @@
+@model Chirp.Authentication.Pages.TimelineModel
+
+@if (Model.Cheeps!.Any())
+{
+    <ul id="messagelist" class="cheeps">
+        @foreach (var cheep in Model.Cheeps!)
+        {
+            <li>
+                <p>
+                    <strong>
+                        <a href="/author/@cheep.Author">@cheep.Author</a>
+                    </strong>
+                    @cheep.Message
+                    <small>&mdash; @cheep.TimeStamp</small>
+                </p>
+            </li>
+        }
+    </ul>
+}
+else
+{
+    <em>There are no cheeps so far.</em>
+}
